Refactored encounter.py to separate class files
