+++ src/model/interfaces/i_tile.py
@@ -6,7 +6,7 @@ game board, including their exits, encounters, and special properties.
 
 from abc import ABC, abstractmethod
 from src.enums_and_types import *
-from ..encounters.encounters import IEncounter
+from ..encounters.not_implemented_encounters import IEncounter
 
 
 class ITile(ABC):
